+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [0.1.87] - 2026-05-19
+
+- Add utility to explicitly set exception on the current active span
+- Update SpanIOProcessor to prioritize input and output attributes explicitly set by user over attributes from instrumentation
+
 ## [0.1.86] - 2026-05-15
 
 - Modify instrument resolution in traceloop to manual transfer of instruments